+++ crates/jsonschema/src/compiler.rs
@@ -327,12 +327,12 @@ pub(crate) fn build_validator(
             .expect("Existing draft")
             .validate(schema)
         {
-            return Err(error.into_owned());
+            return Err(error.to_owned());
         }
     }
 
     // Finally, compile the validator
-    let root = compile(&ctx, resource_ref).map_err(|err| err.into_owned())?;
+    let root = compile(&ctx, resource_ref).map_err(|err| err.to_owned())?;
     Ok(Validator { root, config })
 }
 
@@ -406,7 +406,7 @@ pub(crate) fn compile_with<'a>(
                 } else if let Some((keyword, validator)) = keywords::get_for_draft(ctx, keyword)
                     .and_then(|(keyword, f)| f(ctx, schema, value).map(|v| (keyword, v)))
                 {
-                    validators.push((keyword, validator.map_err(|err| err.into_owned())?));
+                    validators.push((keyword, validator.map_err(|err| err.to_owned())?));
                 } else if !ctx.is_known_keyword(keyword) {
                     // Treat all non-validation keywords as annotations
                     annotations.insert(keyword.to_string(), value.clone());